@@ -9,67 +9,66 @@ import org.matomo.java.tracking.MatomoRequest;
 import org.matomo.java.tracking.MatomoRequest.MatomoRequestBuilder;
 import org.matomo.java.tracking.MatomoTracker;
 import org.matomo.java.tracking.parameters.AcceptLanguage;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
-import org.springframework.util.StringUtils;
 
-import javax.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletRequest;
 import java.util.*;
 
 @Slf4j
 @Component
-public class AsyncMatomoCidResolutionService {
+public class MatomoTrackingService {
     @Value("${org.identifiers.matomo.authToken}")
     String authToken;
 
     @Value("${org.identifiers.matomo.enabled}")
     public boolean isEnabled;
 
-    @Autowired
-    MatomoTracker idorgMatomoTracker;
-
-    private class MatomoRequestImportantInfo {
-        public String url, xforwarded_for, remote_addr, ua, lang, refe;
-        public List<ResolvedResource> resolvedResources;
-        public ParsedCompactIdentifier parsed_cid;
-        public boolean wasResolutionSuccessfull;
-
-        MatomoRequestImportantInfo(final HttpServletRequest request, final ServiceResponse<ResponseResolvePayload> result) {
-            url = request.getRequestURL().toString();
-            xforwarded_for = request.getHeader("X-FORWARDED-FOR");
-            remote_addr = request.getRemoteAddr();
-            ua = request.getHeader("user-agent");
-            lang = request.getHeader("Accept-Language");
-            refe = request.getHeader("Referer");
-
-            ResponseResolvePayload resolveResult = result.getPayload();
-            parsed_cid = resolveResult.getParsedCompactIdentifier();
-            resolvedResources = resolveResult.getResolvedResources();
-
-            wasResolutionSuccessfull = result.getHttpStatus().is2xxSuccessful();
-        }
+    final MatomoTracker idorgMatomoTracker;
+    public MatomoTrackingService(MatomoTracker idorgMatomoTracker) {
+        this.idorgMatomoTracker = idorgMatomoTracker;
     }
 
-    public void asyncHandleCidResolution(final HttpServletRequest request, final ServiceResponse<ResponseResolvePayload> result) {
+    private record MatomoTrackingInfo(
+        String url, String xforwardedFor, String remoteAddr,
+        String ua, String lang, String refe,
+        List<ResolvedResource> resolvedResources,
+        ParsedCompactIdentifier parsedCid,
+        boolean wasResolutionSuccessful
+    ) {}
+
+    private static MatomoTrackingInfo getMatomoInfoFrom(
+            final HttpServletRequest request,
+            final ServiceResponse<ResponseResolvePayload> result
+    ) {
+        return new MatomoTrackingInfo (
+            request.getRequestURL().toString(),
+            request.getHeader("X-FORWARDED-FOR"),
+            request.getRemoteAddr(),
+            request.getHeader("user-agent"),
+            request.getHeader("Accept-Language"),
+            request.getHeader("Referer"),
+            result.getPayload().getResolvedResources(),
+            result.getPayload().getParsedCompactIdentifier(),
+            result.getHttpStatus().is2xxSuccessful());
+    }
+
+    public void handleCidResolution(final HttpServletRequest request, final ServiceResponse<ResponseResolvePayload> result) {
         if (!isEnabled) {
             log.debug("Matomo is disabled");
         } else if (request.getHeader("DNT") != null &&
                 request.getHeader("DNT").contains("1")) {
             log.debug("Skipping matomo notification - DoNotTrack");
         } else {
-            doHandleCidResolution(new MatomoRequestImportantInfo(request, result));
+            doHandleCidResolution(getMatomoInfoFrom(request, result));
         }
     }
 
-    @Async("matomoThreadPoolTaskExecutor")
-    public void doHandleCidResolution(MatomoRequestImportantInfo info) {
+    public void doHandleCidResolution(MatomoTrackingInfo info) {
         MatomoRequestBuilder mreq = MatomoRequest.request();
         mreq.siteId(1);
 
         mreq.actionUrl(info.url);
-//        mreq.trackBotRequests(true);
 
         List<MatomoRequest> requests = new ArrayList<>(2);
 
@@ -78,7 +77,7 @@ public class AsyncMatomoCidResolutionService {
         mreq.newVisit(true);
         requests.add(0, mreq.build());
 
-        if (info.wasResolutionSuccessfull) {
+        if (info.wasResolutionSuccessful) {
             setResolutionResultOnRequest(info, mreq);
             mreq.actionName("Resolution/Redirect");
             mreq.newVisit(false);
@@ -88,16 +87,16 @@ public class AsyncMatomoCidResolutionService {
         idorgMatomoTracker.sendBulkRequestAsync(requests);
     }
 
-    private void setHttpHeadersOnRequest(final MatomoRequestImportantInfo info, MatomoRequestBuilder mreq) {
+    private void setHttpHeadersOnRequest(final MatomoTrackingInfo info, MatomoRequestBuilder mreq) {
         if (authToken != null && !authToken.isEmpty()) {
             mreq.authToken(authToken);
-            String remoteIP = info.xforwarded_for;
+            String remoteIP = info.xforwardedFor;
             if (remoteIP == null || remoteIP.isEmpty()) {
-                remoteIP = info.remote_addr;
+                remoteIP = info.remoteAddr;
             } else { // Make sure to only take client Ip from forwarded for header and discard proxies
                 int firstCommaIdx = remoteIP.indexOf(",");
                 if (firstCommaIdx != -1) {
-                    remoteIP = StringUtils.trimWhitespace(remoteIP.substring(0, firstCommaIdx));
+                    remoteIP = remoteIP.substring(0, firstCommaIdx).strip();
                 }
             }
             mreq.visitorIp(remoteIP);
@@ -109,10 +108,10 @@ public class AsyncMatomoCidResolutionService {
         mreq.referrerUrl(info.refe);
     }
 
-    Comparator<ResolvedResource> resolvedResourceComparator = (o1, o2) ->
-        o1.getRecommendation().getRecommendationIndex() - o2.getRecommendation().getRecommendationIndex();
+    Comparator<ResolvedResource> resolvedResourceComparator =
+            Comparator.comparingInt(o -> o.getRecommendation().getRecommendationIndex());
 
-    private void setResolutionResultOnRequest(MatomoRequestImportantInfo info, MatomoRequestBuilder mreq) {
+    private void setResolutionResultOnRequest(MatomoTrackingInfo info, MatomoRequestBuilder mreq) {
         ResolvedResource maxResolvedResource = info.resolvedResources.stream()
                 .max(resolvedResourceComparator).orElse(null);
         if (maxResolvedResource != null) {
@@ -120,9 +119,9 @@ public class AsyncMatomoCidResolutionService {
         }
 
         Map<Long, Object> customData = new HashMap<>();
-        if (info.parsed_cid != null) {
-            customData.put(6L, info.parsed_cid.getNamespace() == null ? "" : info.parsed_cid.getNamespace());
-            customData.put(7L, info.parsed_cid.getProviderCode() == null ? "" : info.parsed_cid.getProviderCode());
+        if (info.parsedCid != null) {
+            customData.put(6L, info.parsedCid.getNamespace() == null ? "" : info.parsedCid.getNamespace());
+            customData.put(7L, info.parsedCid.getProviderCode() == null ? "" : info.parsedCid.getProviderCode());
         }
         customData.put(8L, info.resolvedResources.size());
         if (maxResolvedResource != null) {
