@@ -2,31 +2,28 @@ package org.identifiers.cloud.ws.resolver.models;
 
 import org.identifiers.cloud.ws.resolver.data.models.Namespace;
 import org.identifiers.cloud.ws.resolver.data.repositories.NamespaceRespository;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
 
 import java.util.Collections;
+import java.util.stream.Stream;
 
-import static org.junit.Assert.*;
+import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.doReturn;
+import static org.mockito.Mockito.mock;
 
-@SpringBootTest
-@RunWith(SpringRunner.class)
-public class CompactIdParsingHelperTest {
+class CompactIdParsingHelperTest {
+    final CompactIdParsingHelper helper;
+    final NamespaceRespository namespaceRespository;
+    public CompactIdParsingHelperTest() {
+        namespaceRespository = mock(NamespaceRespository.class);
+        helper = new CompactIdParsingHelper(namespaceRespository);
+    }
 
-    @Autowired
-    CompactIdParsingHelper helper;
-
-    @MockBean
-    NamespaceRespository namespaceRespository;
-
-    @Before
+    @BeforeEach
     public void setupNamespaceRepository() {
         Namespace namespace = new Namespace()
                 .setPrefix("existing_namespace")
@@ -44,74 +41,56 @@ public class CompactIdParsingHelperTest {
         doReturn(luinamespace).when(namespaceRespository).findByPrefix("lui_namespace");
     }
 
-    @Test
-    public void parseCompactIdRequests() {
-        testRawRequest("Namespace only",
-                "existing_namespace",
-                null, "existing_namespace", null);
+    private static Stream<Arguments> CompactIdParsingHelperTestParameters() {
+        return Stream.of(
+            ////// Tests with existing namespace
+            Arguments.of("Namespace only",
+                    "existing_namespace", null, "existing_namespace", null),
+            Arguments.of("Namespace + lui",
+                    "existing_namespace:123", null, "existing_namespace", "123"),
+            Arguments.of("Namespace + invalid lui",
+                    "existing_namespace:abc", null, "existing_namespace", "abc"),
+            Arguments.of("Namespace + lui using slash",
+                    "existing_namespace/123", null, "existing_namespace", "123"),
+            Arguments.of("Namespace + lui + provider",
+                    "provider/existing_namespace:123", "provider", "existing_namespace", "123"),
+            //FIXME: The parsed result below should be the same as the previous
+            Arguments.of("Namespace + lui + provider using slash",
+                    "provider/existing_namespace/123", "provider", null, null),
 
-        testRawRequest("Namespace w/lui",
-                "existing_namespace:123",
-                null, "existing_namespace", "123");
-        testRawRequest("Namespace w/ invalid lui",
-                "existing_namespace:abc",
-                null, "existing_namespace", "abc");
-        testRawRequest("Namespace w/lui using slash",
-                "existing_namespace/123",
-                null, "existing_namespace", "123");
+            ////// Tests with non existing namespace
+            // FIXME: The helper is inconsistent when the namespace doesn't exist
+            Arguments.of("Non existing namespace",
+                    "non_existing_namespace", null, null, null),
+            Arguments.of("Non existing namespace + lui",
+                    "non_existing_namespace:123", null, "non_existing_namespace", "123"),
+            Arguments.of("Non existing namespace + lui using slash",
+                    "non_existing_namespace/123", "non_existing_namespace", null, null),
+            Arguments.of("Non existing namespace + lui + provider",
+                    "provider/non_existing_namespace:123", "provider", null, null),
+            Arguments.of("Non existing namespace + lui + provider using slash",
+                    "provider/non_existing_namespace/123", "provider", null, null),
 
-        testRawRequest("Namespace w/lui and provider code",
-                "provider/existing_namespace:123",
-                "provider", "existing_namespace", "123");
-
-        //FIXME: The parsed result below should be the same as the previous
-        testRawRequest("Namespace w/lui and provider code using slash",
-                "provider/existing_namespace/123",
-                "provider", null, null);
+            ////// Tests with embedded lui namespace
+            Arguments.of("Embedded lui namespace + lui ",
+                    "lui_namespace:123", null, "lui_namespace", "lui_namespace:123"),
+            Arguments.of("Embedded lui namespace + lui + provider",
+                    "provider/lui_namespace:123", "provider", "lui_namespace", "lui_namespace:123"),
+            //FIXME: The below is inconsistent
+            Arguments.of("Embedded lui namespace + lui + provider using slash",
+                    "provider/lui_namespace/123", "provider", null, null)
+        );
     }
 
-    @Test
-    public void parseCompactIdRequestsWithNonExistingNamespace() {
-        // FIXME: The helper is inconsistent when the namespace doesn't exist
-        testRawRequest("Non existing namespace only",
-                "non_existing_namespace",
-                null, null, null);
-
-        testRawRequest("Non existing namespace w/lui",
-                "non_existing_namespace:123",
-                null, "non_existing_namespace", "123");
-        testRawRequest("Non existing namespace w/lui using slash",
-                "non_existing_namespace/123",
-                "non_existing_namespace", null, null);
-
-        testRawRequest("Non existing namespace w/lui using slash",
-                "provider/non_existing_namespace:123",
-                "provider", null, null);
-        testRawRequest("Non existing namespace w/lui using slash",
-                "provider/non_existing_namespace/123",
-                "provider", null, null);
-    }
-
-    @Test
-    public void parseCompactIdRequestsWithLuiNamespace() {
-        testRawRequest("Namespace w/lui",
-                "lui_namespace:123",
-                null, "lui_namespace", "lui_namespace:123");
-        testRawRequest("Namespace w/lui and provider code",
-                "provider/lui_namespace:123",
-                "provider", "lui_namespace", "lui_namespace:123");
-
-        //FIXME: The below is inconsistent
-        testRawRequest("Namespace w/lui and provider code",
-                "provider/lui_namespace/123",
-                "provider", null, null);
-    }
-
-    void testRawRequest(String msg, String rawRequest, String providerCode, String namespace, String localId) {
-        ParsedCompactIdentifier pci = helper.parseCompactIdRequest(rawRequest);
-        assertEquals(msg, rawRequest, pci.getRawRequest());
-        assertEquals(msg, providerCode, pci.getProviderCode());
-        assertEquals(msg, namespace, pci.getNamespace());
-        assertEquals(msg, localId, pci.getLocalId());
+    @ParameterizedTest(name = "{0}")
+    @MethodSource("CompactIdParsingHelperTestParameters")
+    void testRawRequest(String msg,
+                        String expectedRawRequest, String expectedProviderCode,
+                        String expectedNamespace, String expectedLocalId) {
+        ParsedCompactIdentifier pci = helper.parseCompactIdRequest(expectedRawRequest);
+        assertEquals(expectedRawRequest, pci.getRawRequest(), msg);
+        assertEquals(expectedProviderCode, pci.getProviderCode(), msg);
+        assertEquals(expectedNamespace, pci.getNamespace(), msg);
+        assertEquals(expectedLocalId, pci.getLocalId(), msg);
     }
 }
