@@ -1,48 +1,41 @@
-package org.identifiers.cloud.ws.resolver.daemons;
+package org.identifiers.cloud.ws.resolver.periodictasks;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.Slf4j;
-import org.identifiers.cloud.ws.resolver.RedisTestServer;
-import org.identifiers.cloud.ws.resolver.daemons.models.HqServiceResponseGetResolverDataset;
-import org.identifiers.cloud.ws.resolver.daemons.models.ResolverDatasetPayload;
-import org.identifiers.cloud.ws.resolver.daemons.models.ServiceResponse;
+import org.identifiers.cloud.ws.resolver.TestRedisServer;
+import org.identifiers.cloud.ws.resolver.periodictasks.models.HqServiceResponseGetResolverDataset;
+import org.identifiers.cloud.ws.resolver.periodictasks.models.ResolverDatasetPayload;
+import org.identifiers.cloud.ws.resolver.periodictasks.models.ServiceResponse;
 import org.identifiers.cloud.ws.resolver.data.models.Namespace;
 import org.identifiers.cloud.ws.resolver.data.models.Resource;
 import org.identifiers.cloud.ws.resolver.data.repositories.NamespaceRespository;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.client.MockRestServiceServer;
 import org.springframework.web.client.RestTemplate;
-
-import javax.annotation.PostConstruct;
 
 import java.util.Collections;
 import java.util.List;
 
-import static org.junit.Assert.*;
-import static org.springframework.test.web.client.ExpectedCount.manyTimes;
+import static org.junit.jupiter.api.Assertions.*;
+import static org.springframework.test.web.client.ExpectedCount.once;
 import static org.springframework.test.web.client.match.MockRestRequestMatchers.requestTo;
 import static org.springframework.test.web.client.response.MockRestResponseCreators.withStatus;
 
-@SpringBootTest(classes = RedisTestServer.class)
-@RunWith(SpringRunner.class)
 @Slf4j
-public class ResolverDataUpdaterTest {
+@SpringBootTest(classes = TestRedisServer.class)
+class ResolverDataUpdaterTest {
 
     @Autowired
     NamespaceRespository namespaceRespository;
-
     @Autowired
     ResolverDataUpdater updater;
-
     @Autowired
     RestTemplate restTemplate;
 
@@ -51,12 +44,7 @@ public class ResolverDataUpdaterTest {
 
     final ObjectMapper mapper = new ObjectMapper();
 
-    @PostConstruct
-    public void stopUpdaterThread() {
-        updater.setShutdown();
-    }
-
-    @Before
+    @BeforeEach
     public void setupRegistryServer() throws JsonProcessingException {
         List<Namespace> mockedNamespaces = Collections.singletonList(
                 new Namespace().setPrefix("mock_prefix").setPattern("\\d+").setName("Mocked Prefix")
@@ -70,15 +58,15 @@ public class ResolverDataUpdaterTest {
         String responseStr = mapper.writeValueAsString(response);
 
         MockRestServiceServer server = MockRestServiceServer.createServer(restTemplate);
-        server.expect(manyTimes(), requestTo(resolverDataDumpWsEndpoint))
+        server.expect(once(), requestTo(resolverDataDumpWsEndpoint))
                 .andRespond(withStatus(HttpStatus.OK).body(responseStr).contentType(MediaType.APPLICATION_JSON));
     }
 
     @Test
-    public void updateNamespace() { //FIXME: this only works when run by itself. Probably move redis server to test configuration
-        updater.updateNamespaces();
-        assertEquals("Only find one namespace", 1, namespaceRespository.count());
-        assertNotNull("Only find mocked prefix namespace", namespaceRespository.findByPrefix("mock_prefix"));
-        assertEquals("Only one resource", 1, namespaceRespository.findByPrefix("mock_prefix").getResources().size());
+    void updateNamespace() {
+        updater.run();
+        assertEquals(1, namespaceRespository.count());
+        assertNotNull(namespaceRespository.findByPrefix("mock_prefix"));
+        assertEquals(1, namespaceRespository.findByPrefix("mock_prefix").getResources().size());
     }
 }
