@@ -1,28 +1,27 @@
 package org.identifiers.cloud.ws.resolver.services;
 
+import org.identifiers.cloud.ws.resolver.TestRedisServer;
 import org.identifiers.cloud.ws.resolver.data.models.Location;
 import org.identifiers.cloud.ws.resolver.data.models.Namespace;
 import org.identifiers.cloud.ws.resolver.data.models.Resource;
 import org.identifiers.cloud.ws.resolver.data.repositories.NamespaceRespository;
 import org.identifiers.cloud.ws.resolver.models.CompactIdParsingHelper;
 import org.identifiers.cloud.ws.resolver.models.ParsedCompactIdentifier;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.test.context.junit4.SpringRunner;
 
 import java.util.Collections;
 
-import static org.junit.Assert.*;
+import static org.junit.jupiter.api.Assertions.*;
+
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.doReturn;
 
-@SpringBootTest
-@RunWith(SpringRunner.class)
-public class ResolutionServiceTest {
+@SpringBootTest(classes = {TestRedisServer.class})
+class ResolutionServiceTest {
     @Autowired ResolutionService resolutionService;
     @Autowired CompactIdParsingHelper helper;
 
@@ -31,7 +30,7 @@ public class ResolutionServiceTest {
 
 
 
-    @Before
+    @BeforeEach
     public void setupNamespaceRepository() {
         Namespace namespace = new Namespace()
                 .setPrefix("existing_namespace")
@@ -48,41 +47,38 @@ public class ResolutionServiceTest {
     }
 
     @Test
-    public void testResolveExistingNamespace() {
+    void testResolveExistingNamespace() {
         ParsedCompactIdentifier pci = helper.parseCompactIdRequest("existing_namespace");
         ResolutionServiceResult result = resolutionService.resolve(pci);
-        assertTrue("Finds valid namespace", result.isResolved());
-        assertEquals("Only idorg is found as resource for a namespace",
-                1, result.getResolvedResources().size());
-        assertNull("Resolved namespace should be null",
-                result.getResolvedResources().get(0).getNamespacePrefix());
+        assertTrue(result.isResolved());
+        assertEquals(1, result.getResolvedResources().size());
+        assertNull(result.getResolvedResources().get(0).getNamespacePrefix());
     }
 
     @Test
-    public void testResolveNonExistingNamespace() {
+    void testResolveNonExistingNamespace() {
         ParsedCompactIdentifier pci = helper.parseCompactIdRequest("non_namespace");
         ResolutionServiceResult result = resolutionService.resolve(pci);
-        assertFalse("Finds valid namespace", result.isResolved());
-        assertEquals("Only idorg is found as resource for a namespace",
-                0, result.getResolvedResources().size());
+        assertFalse(result.isResolved());
+        assertEquals(0, result.getResolvedResources().size());
     }
 
     @Test
-    public void testResolveValidCID() {
+    void testResolveValidCID() {
         ParsedCompactIdentifier pci = helper.parseCompactIdRequest("existing_namespace:123");
         ResolutionServiceResult result = resolutionService.resolve(pci);
-        assertTrue("Finds valid namespace", result.isResolved());
-        assertEquals("only one resource is found for it",1, result.getResolvedResources().size());
+        assertTrue(result.isResolved());
+        assertEquals(1, result.getResolvedResources().size());
         String resolvedURL = result.getResolvedResources().get(0).getCompactIdentifierResolvedUrl();
-        assertTrue("Resolved resource is mocked", resolvedURL.startsWith("https://mockedurl.com"));
-        assertFalse("Resolved URL should not contain template variable", resolvedURL.contains("{$id}"));
+        assertTrue(resolvedURL.startsWith("https://mockedurl.com"));
+        assertFalse(resolvedURL.contains("{$id}"));
     }
 
     @Test
-    public void testResolveNonValidCID() {
+    void testResolveNonValidCID() {
         ParsedCompactIdentifier pci = helper.parseCompactIdRequest("existing_namespace:abc");
         ResolutionServiceResult result = resolutionService.resolve(pci);
-        assertFalse("Should not resolve invalid local id", result.isResolved());
-        assertTrue("Empty resource for invalid local id", result.getResolvedResources().isEmpty());
+        assertFalse(result.isResolved());
+        assertTrue(result.getResolvedResources().isEmpty());
     }
 }
