@@ -13,7 +13,7 @@ import java.util.List;
 public class CompactId {
     public static final String COMPACT_ID_PREFIX_AND_ID_SEPARATOR = ":";
 
-    private String original;
+    private final String original;
     private String prefix = null;
     private String id = null;
 
